+++ models/album.js
@@ -3,7 +3,7 @@ const Schema   = mongoose.Schema;
 
 const albumSchema = new Schema({
   name: String,
-  userId: { type : Schema.Types.ObjectId, ref: 'User' }
+  pictureID: [{ type : Schema.Types.ObjectId, ref: 'Picture' }]
 }, {
   timestamps: { createdAt: "created_at", updatedAt: "updated_at" }
 });
