+++ views/partials/cards/picture.hbs
@@ -2,6 +2,20 @@
 
 {{photo}}
 
+<div class="photo">
+  <div class="photo__actions">
+    <div class="photo__actions-option">
+      <a href="#"><i class="fas fa-pencil-alt"></i></a>
+    </div>
+    <div class="photo__actions-option">
+      <a href="#"><i class="fas fa-trash"></i></a>
+    </div>
+  </div>
+  <a href="/mock/photo">
+    <img src="/images/test/test-2.jpg" alt="Test 1" class="photo__image">
+  </a>
+</div>
+
 {{!-- <form action="/pict/newPicture" method="post" enctype="multipart/form-data">
   <label>Title</label>
   <input type="text" name="title">
